+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 ### Added
 * Add Roles and Permissions Support
+* Add Headers, Topics and Webhook Name Support
 
 ### Changed
 * Added Deprecation warning for `assets.all_published` and `entries.all_published`. This methods will be completely removed soon.